Pybind11 update now with setup.py
Updated the pybind11 interface to include some function calls and also
added a setup.py. This appears to work but significant work to be done
to add full functionality.

Need to see how best to process the git_revision.cpp.in in the setup.py
so that it gets copied, updated, compiled removed and also
add in the cmake interface for the pybind11 build.